Major Refactoring - Tests do not parse (yet)
Realized I tried to retain too much knowledge about the parts in play in the ByteSchema itself. Which will make it either unmaintainable when I get to more abstract patterns. A "tree"-structure of accumulation and parsing is probably a better way to go to keep the complexity of individual elements low